+++ cider.el
@@ -709,13 +709,27 @@ Generally you should not disable this unless you run into some faulty check."
   :safe (lambda (s) (or (null s) (stringp s)))
   :package-version '(cider . "0.18.0"))
 
+(defun cider--shadow-parse-builds ()
+  "Extract build names from the shadow-cljs.edn config file in the project root.
+The default options of `browser-repl' and `node-repl' are also included."
+  (let* ((shadow-edn
+          (concat (clojure-project-dir (cider-current-dir)) "shadow-cljs.edn"))
+         (edn-hash (when (file-exists-p shadow-edn)
+                     (with-temp-buffer
+                       (insert-file-contents shadow-edn)
+                       (car (parseedn-read)))))
+         (builds-hash (when edn-hash (gethash :builds edn-hash)))
+         (builds (when builds-hash (hash-table-keys builds-hash))))
+    (append builds '(browser-repl node-repl))))
+
 (defun cider-shadow-select-cljs-init-form ()
   "Generate the init form for a shadow-cljs select-only REPL.
 We have to prompt the user to select a build, that's why this is a command,
 not just a string."
   (let ((form "(do (require '[shadow.cljs.devtools.api :as shadow]) (shadow/nrepl-select %s))")
         (options (or cider-shadow-default-options
-                     (read-from-minibuffer "Select shadow-cljs build (e.g. dev): "))))
+                     (completing-read "Select shadow-cljs build: "
+                                      (cider--shadow-parse-builds)))))
     (format form (cider-normalize-cljs-init-options options))))
 
 (defun cider-shadow-cljs-init-form ()
@@ -724,7 +738,8 @@ We have to prompt the user to select a build, that's why
 this is a command, not just a string."
   (let* ((form "(do (require '[shadow.cljs.devtools.api :as shadow]) (shadow/watch %s) (shadow/nrepl-select %s))")
          (options (or cider-shadow-default-options
-                      (read-from-minibuffer "Select shadow-cljs build (e.g. dev): ")))
+                      (completing-read "Select shadow-cljs build: "
+                                       (cider--shadow-parse-builds))))
          (build (cider-normalize-cljs-init-options options)))
     (format form build build)))